requirements.txt
# BSK Service Recommendation System - Requirements

## Core Dependencies
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.28.0

## Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0  # For PostgreSQL
python-dotenv>=0.19.0

## AI/ML Dependencies  
sentence-transformers>=2.2.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.21.0

## Vector Database
chromadb>=0.4.0

## Geospatial
geopy>=2.3.0

## NLP
nltk>=3.8.0

## Visualization
pydeck>=0.8.0

## Backend API
fastapi>=0.95.0
uvicorn>=0.20.0

## Optional Performance
numba>=0.57.0  # For faster computations
